CLN: simplify _shallow_copy
