+++ pandas/core/base.py
@@ -634,20 +634,19 @@ class SelectionMixin:
 class ShallowMixin:
     _attributes = []  # type: List[str]
 
-    def _shallow_copy(self, obj=None, obj_type=None, **kwargs):
+    def _shallow_copy(self, obj=None, **kwargs):
         """
         return a new object with the replacement attributes
         """
         if obj is None:
             obj = self._selected_obj.copy()
-        if obj_type is None:
-            obj_type = self._constructor
-        if isinstance(obj, obj_type):
+
+        if isinstance(obj, self._constructor):
             obj = obj.obj
         for attr in self._attributes:
             if attr not in kwargs:
                 kwargs[attr] = getattr(self, attr)
-        return obj_type(obj, **kwargs)
+        return self._constructor(obj, **kwargs)
 
 
 class IndexOpsMixin:
